Collect cpp and hpp files to a single directory
* Separate files of binary and shared object from the library code.
* Separate the CMakeLists.txt file list variables in the directory hierarchy.
