+++ cpp/modmesh/mesh/StaticMesh_interior.cpp
