+++ cpp/modmesh/view/PythonInterpreter.cpp
@@ -26,13 +26,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/viewer/base.hpp> // Must be the first include.
+#include <modmesh/view/base.hpp> // Must be the first include.
 
-#include <modmesh/viewer/PythonInterpreter.hpp>
+#include <modmesh/view/PythonInterpreter.hpp>
 
-#include <modmesh/viewer/RMainWindow.hpp>
-#include <modmesh/viewer/RStaticMesh.hpp>
-#include <modmesh/viewer/R3DWidget.hpp>
+#include <modmesh/view/RMainWindow.hpp>
+#include <modmesh/view/RStaticMesh.hpp>
+#include <modmesh/view/R3DWidget.hpp>
 
 namespace modmesh
 {
