+++ cpp/modmesh/view/R3DWidget.cpp
@@ -26,8 +26,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/viewer/base.hpp> // Must be the first include.
-#include <modmesh/viewer/R3DWidget.hpp>
+#include <modmesh/view/base.hpp> // Must be the first include.
+#include <modmesh/view/R3DWidget.hpp>
 
 namespace modmesh
 {
