+++ cpp/modmesh/view/RMainWindow.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 /*
  * Copyright (c) 2022, Yung-Yu Chen <yyc@solvcon.net>
  *
@@ -28,40 +26,34 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/viewer/base.hpp> // Must be the first include.
+#include <modmesh/view/base.hpp> // Must be the first include.
+#include <modmesh/view/RMainWindow.hpp>
 
-#include <Qt>
-#include <QDockWidget>
-#include <QTextEdit>
-#include <QPushButton>
-#include <QVBoxLayout>
+#include <modmesh/view/PythonInterpreter.hpp>
+#include <modmesh/view/RPythonText.hpp>
+#include <modmesh/view/R3DWidget.hpp>
 
 namespace modmesh
 {
 
-class RPythonText
-    : public QDockWidget
+void RMainWindow::setUp()
 {
+    m_pytext = new RPythonText(QString("Python"), this);
+    m_pytext->setAllowedAreas(Qt::LeftDockWidgetArea | Qt::RightDockWidgetArea);
+    addDockWidget(Qt::RightDockWidgetArea, m_pytext);
 
-public:
+    m_viewer = new R3DWidget();
+    setCentralWidget(m_viewer);
+}
 
-    RPythonText(
-        QString const & title = "Python",
-        QWidget * parent = nullptr,
-        Qt::WindowFlags flags = Qt::WindowFlags());
-
-private:
-
-    void setUp();
-
-    void runPythonCode();
-
-    QTextEdit * m_text = nullptr;
-    QPushButton * m_run = nullptr;
-    QVBoxLayout * m_layout = nullptr;
-    QWidget * m_widget = nullptr;
-
-}; /* end class RPythonText */
+RApplication::RApplication(int & argc, char ** argv)
+    : QApplication(argc, argv)
+    , m_main(new RMainWindow)
+{
+    /* TODO: parse arguments */
+    PythonInterpreter::instance();
+    m_main->show();
+}
 
 } /* end namespace modmesh */
